moved course definition to courseNode
